+++ pkg/pyepo/func/blackbox.py
@@ -4,22 +4,16 @@
 Differentiable Black-box optimization function
 """
 
-import multiprocessing as mp
-
 import numpy as np
 import torch
-from pathos.multiprocessing import ProcessingPool
 from torch.autograd import Function
-from torch import nn
 
+from pyepo.func.abcmodule import optModule
 from pyepo import EPO
-from pyepo.data.dataset import optDataset
-from pyepo.model.opt import optModel
-
 from pyepo.func.utlis import _solveWithObj4Par, _solve_in_pass, _cache_in_pass
 
 
-class blackboxOpt(nn.Module):
+class blackboxOpt(optModule):
     """
     A autograd module for differentiable black-box optimizer, which yield
     optimal a solution and derive a gradient.
@@ -41,37 +35,11 @@ class blackboxOpt(nn.Module):
             solve_ratio (float): the ratio of new solutions computed during training
             dataset (None/optDataset): the training data
         """
-        super().__init__()
-        # optimization model
-        if not isinstance(optmodel, optModel):
-            raise TypeError("arg model is not an optModel.")
-        self.optmodel = optmodel
+        super().__init__(optmodel, processes, solve_ratio, dataset)
         # smoothing parameter
         if lambd <= 0:
             raise ValueError("lambda is not positive.")
         self.lambd = lambd
-        # number of processes
-        if processes not in range(mp.cpu_count()+1):
-            raise ValueError("Invalid processors number {}, only {} cores.".
-                format(processes, mp.cpu_count()))
-        self.processes = mp.cpu_count() if not processes else processes
-        # single-core
-        if processes == 1:
-            self.pool = None
-        # multi-core
-        else:
-            self.pool = ProcessingPool(processes)
-        print("Num of cores: {}".format(self.processes))
-        # solution pool
-        self.solve_ratio = solve_ratio
-        if (self.solve_ratio < 0) or (self.solve_ratio > 1):
-            raise ValueError("Invalid solving ratio {}. It should be between 0 and 1.".
-                format(self.solve_ratio))
-        self.solpool = None
-        if self.solve_ratio < 1: # init solution pool
-            if not isinstance(dataset, optDataset): # type checking
-                raise TypeError("dataset is not an optDataset")
-            self.solpool = dataset.sols.copy()
         # build blackbox optimizer
         self.dbb = blackboxOptFunc()
 
@@ -101,7 +69,7 @@ class blackboxOptFunc(Function):
             processes (int): number of processors, 1 for single-core, 0 for all of cores
             pool (ProcessPool): process pool object
             solve_ratio (float): the ratio of new solutions computed during training
-            module (nn.Module): blackboxOpt module
+            module (optModule): blackboxOpt module
 
         Returns:
             torch.tensor: predicted solutions
