+++ pkg/pyepo/func/perturbed.py
@@ -4,21 +4,16 @@
 Perturbed optimization function
 """
 
-import multiprocessing as mp
-
 import numpy as np
 import torch
-from pathos.multiprocessing import ProcessingPool
 from torch.autograd import Function
-from torch import nn
 
 from pyepo import EPO
-from pyepo.data.dataset import optDataset
-from pyepo.model.opt import optModel
+from pyepo.func.abcmodule import optModule
 from pyepo.utlis import getArgs
 
 
-class perturbedOpt(nn.Module):
+class perturbedOpt(optModule):
     """
     A autograd module for differentiable perturbed optimizer, in which random
     perturbed costs are sampled to optimize.
@@ -42,39 +37,13 @@ class perturbedOpt(nn.Module):
             solve_ratio (float): the ratio of new solutions computed during training
             dataset (None/optDataset): the training data
         """
-        super().__init__()
-        # optimization model
-        if not isinstance(optmodel, optModel):
-            raise TypeError("arg model is not an optModel")
-        self.optmodel = optmodel
+        super().__init__(optmodel, processes, solve_ratio, dataset)
         # number of samples
         self.n_samples = n_samples
         # perturbation amplitude
         self.sigma = sigma
-        # number of processes
-        if processes not in range(mp.cpu_count()+1):
-            raise ValueError("Invalid processors number {}, only {} cores.".
-                format(processes, mp.cpu_count()))
-        self.processes = mp.cpu_count() if not processes else processes
-        # single-core
-        if processes == 1:
-            self.pool = None
-        # multi-core
-        else:
-            self.pool = ProcessingPool(processes)
-        print("Num of cores: {}".format(self.processes))
         # random state
         self.rnd = np.random.RandomState(seed)
-        # solution pool
-        self.solve_ratio = solve_ratio
-        if (self.solve_ratio < 0) or (self.solve_ratio > 1):
-            raise ValueError("Invalid solving ratio {}. It should be between 0 and 1.".
-                format(self.solve_ratio))
-        self.solpool = None
-        if self.solve_ratio < 1: # init solution pool
-            if not isinstance(dataset, optDataset): # type checking
-                raise TypeError("dataset is not an optDataset")
-            self.solpool = dataset.sols.copy()
         # build optimizer
         self.ptb = perturbedOptFunc()
 
@@ -108,7 +77,7 @@ class perturbedOptFunc(Function):
             pool (ProcessPool): process pool object
             rnd (RondomState): numpy random state
             solve_ratio (float): the ratio of new solutions computed during training
-            module (nn.Module): perturbedOpt module
+            module (optModule): perturbedOpt module
 
         Returns:
             torch.tensor: solution expectations with perturbation
@@ -159,7 +128,7 @@ class perturbedOptFunc(Function):
         return grad, None, None, None, None, None, None, None, None
 
 
-class perturbedFenchelYoung(nn.Module):
+class perturbedFenchelYoung(optModule):
     """
     A autograd module for Fenchel-Young loss using perturbation techniques. The
     use of the loss improves the algorithmic by the specific expression of the
@@ -185,39 +154,13 @@ class perturbedFenchelYoung(nn.Module):
             solve_ratio (float): the ratio of new solutions computed during training
             dataset (None/optDataset): the training data
         """
-        super().__init__()
-        # optimization model
-        if not isinstance(optmodel, optModel):
-            raise TypeError("arg model is not an optModel")
-        self.optmodel = optmodel
+        super().__init__(optmodel, processes, solve_ratio, dataset)
         # number of samples
         self.n_samples = n_samples
         # perturbation amplitude
         self.sigma = sigma
-        # number of processes
-        if processes not in range(mp.cpu_count()+1):
-            raise ValueError("Invalid processors number {}, only {} cores.".
-                format(processes, mp.cpu_count()))
-        self.processes = mp.cpu_count() if not processes else processes
-        # single-core
-        if processes == 1:
-            self.pool = None
-        # multi-core
-        else:
-            self.pool = ProcessingPool(processes)
-        print("Num of cores: {}".format(self.processes))
         # random state
         self.rnd = np.random.RandomState(seed)
-        # solution pool
-        self.solve_ratio = solve_ratio
-        if (self.solve_ratio < 0) or (self.solve_ratio > 1):
-            raise ValueError("Invalid solving ratio {}. It should be between 0 and 1.".
-                format(self.solve_ratio))
-        self.solpool = None
-        if self.solve_ratio < 1: # init solution pool
-            if not isinstance(dataset, optDataset): # type checking
-                raise TypeError("dataset is not an optDataset")
-            self.solpool = dataset.sols.copy()
         # build optimizer
         self.pfy = perturbedFenchelYoungFunc()
 
@@ -261,7 +204,7 @@ class perturbedFenchelYoungFunc(Function):
             pool (ProcessPool): process pool object
             rnd (RondomState): numpy random state
             solve_ratio (float): the ratio of new solutions computed during training
-            module (nn.Module): perturbedFenchelYoung module
+            module (optModule): perturbedFenchelYoung module
 
         Returns:
             torch.tensor: solution expectations with perturbation
