+++ pkg/pyepo/func/spoplus.py
@@ -4,21 +4,15 @@
 SPO+ Loss function
 """
 
-import multiprocessing as mp
-
 import numpy as np
 import torch
-from pathos.multiprocessing import ProcessingPool
 from torch.autograd import Function
-from torch import nn
 
 from pyepo import EPO
-from pyepo.data.dataset import optDataset
-from pyepo.model.opt import optModel
-
+from pyepo.func.abcmodule import optModule
 from pyepo.func.utlis import _solveWithObj4Par, _solve_in_pass, _cache_in_pass
 
-class SPOPlus(nn.Module):
+class SPOPlus(optModule):
     """
     A autograd module for SPO+ Loss, as a surrogate loss function of SPO Loss,
     which measures the decision error of optimization problem.
@@ -39,33 +33,7 @@ class SPOPlus(nn.Module):
             solve_ratio (float): the ratio of new solutions computed during training
             dataset (None/optDataset): the training data
         """
-        super().__init__()
-        # optimization model
-        if not isinstance(optmodel, optModel):
-            raise TypeError("arg model is not an optModel")
-        self.optmodel = optmodel
-        # number of processes
-        if processes not in range(mp.cpu_count()+1):
-            raise ValueError("Invalid processors number {}, only {} cores.".
-                format(processes, mp.cpu_count()))
-        self.processes = mp.cpu_count() if not processes else processes
-        # single-core
-        if processes == 1:
-            self.pool = None
-        # multi-core
-        else:
-            self.pool = ProcessingPool(processes)
-        print("Num of cores: {}".format(self.processes))
-        # solution pool
-        self.solve_ratio = solve_ratio
-        if (self.solve_ratio < 0) or (self.solve_ratio > 1):
-            raise ValueError("Invalid solving ratio {}. It should be between 0 and 1.".
-                format(self.solve_ratio))
-        self.solpool = None
-        if self.solve_ratio < 1: # init solution pool
-            if not isinstance(dataset, optDataset): # type checking
-                raise TypeError("dataset is not an optDataset")
-            self.solpool = dataset.sols.copy()
+        super().__init__(optmodel, processes, solve_ratio, dataset)
         # build carterion
         self.spop = SPOPlusFunc()
 
@@ -109,7 +77,7 @@ class SPOPlusFunc(Function):
             processes (int): number of processors, 1 for single-core, 0 for all of cores
             pool (ProcessPool): process pool object
             solve_ratio (float): the ratio of new solutions computed during training
-            module (nn.Module): SPOPlus modeul
+            module (optModule): SPOPlus modeul
 
         Returns:
             torch.tensor: SPO+ loss
